move windows code to file excluded from coverage computing
